should overwrite eventual consistency methods
